feat: restructure as monorepo with pnpm workspaces
Extract shared logic (Firebase connection, model parsing, Zod schema
generation) into @contedra/core package. The loader package now imports
from core, keeping the public API unchanged.

Structure:
- packages/core: Firebase init, model JSON reader, schema generator
- packages/astro-loader-firestore: Astro Content Layer loader
- demo: unchanged, references workspace packages

Closes #5
