+++ packages/astro-loader-firestore/src/index.ts
@@ -1,0 +1,3 @@
+export { contedraLoader } from "./loader.js";
+export type { ContedraLoaderConfig } from "./types.js";
+export type { ModelDefinition, ModelProperty } from "@contedra/core";
