+++ packages/core/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "@contedra/core",
+  "version": "0.1.0",
+  "description": "Core library for Contedra toolkit — Firebase connection, model parsing, and schema generation",
+  "type": "module",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    }
+  },
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsc",
+    "test": "vitest run",
+    "lint": "oxlint src"
+  },
+  "keywords": [
+    "conteditor",
+    "firestore",
+    "firebase"
+  ],
+  "author": "",
+  "license": "MIT",
+  "dependencies": {
+    "firebase-admin": "^13.0.0",
+    "zod": "^3.0.0"
+  },
+  "devDependencies": {
+    "oxlint": "^1.57.0",
+    "typescript": "^5.7.0",
+    "vitest": "^3.0.0"
+  }
+}
