+++ packages/core/src/__tests__/firestore.test.ts
