+++ packages/core/src/__tests__/fixtures/blog_posts.json
