+++ packages/core/src/__tests__/model.test.ts
