+++ packages/core/src/__tests__/schema.test.ts
