+++ packages/core/src/index.ts
@@ -1,0 +1,4 @@
+export { loadModel, detectBodyField } from "./model.js";
+export { buildSchema, dataTypeToZod } from "./schema.js";
+export { initFirestore, fetchDocuments, transformDocumentData } from "./firestore.js";
+export type { ModelDefinition, ModelProperty, FirebaseConfig } from "./types.js";
