+++ packages/core/src/model.ts
