+++ packages/core/vitest.config.ts
@@ -3,6 +3,5 @@ import { defineConfig } from "vitest/config";
 export default defineConfig({
   test: {
     globals: true,
-    exclude: ["node_modules/**", "demo/**"],
   },
 });
