--- src/index.ts
@@ -1,2 +1,0 @@
-export { contedraLoader } from "./loader.js";
-export type { ContedraLoaderConfig, ModelDefinition, ModelProperty } from "./types.js";
